fix(agent-plugin): address the second CodeRabbit review
- The publish job reconciles the version tag when the mirror has nothing to
  commit. A previous run can land the branch push and then fail the tag push,
  which left the released content under no version at all while the workflow
  reported success. The tag is now created when it is missing, and verified
  against the published commit when it is there.
- Split the plugin suite along what it exercises: reading state, installing,
  and removing. The setup they shared moves into the fixture module as one
  harness, so the split costs no duplicated preamble beyond the per-file
  spawnSync mock, which vitest hoists and cannot be registered elsewhere.
- Named object parameters for the last positional test helpers.
- The scenarios added for the review fixes read as behaviour: a clock that
  disagrees with the recorded failure, a second logout leaving capture off,
  a marketplace that only mentions our repository.
